Never store precompiled preambles in memory to work around a crash
Precompiled preambles are stored in RAM by default given libclang
version 17 or later, where the necessary API is available since
https://github.com/llvm/llvm-project/commit/55f7e00afc56c68421220d60d29c079b58fe9c79

As discovered in a comment thread starting at
https://invent.kde.org/kdevelop/kdevelop/-/merge_requests/777#note_1224253
the linked crash that affects many KDevelop users can be worked around
by not storing precompiled preambles in memory.

So hide the precompiled preamble storage configuration UI and never
store the preambles in memory until a proper fix is found.

BUG: 486932
FIXED-IN: 6.3.250800