ToolViewAction: use inactive text color instead of parenthesizing
The Frame Stack tool view widget is disabled whenever there is no debug
session or the state of the current debug session is neither paused nor
stopped. The text of a tool button is enclosed in parentheses when its
associated tool view widget is disabled.

The state of the current debug session often alternates between paused
and active during debugging. Thus the tool button text "Frame Stack" is
repeatedly (un)parenthesized. This changes the width of the tool button,
shifts all buttons and the status bar messages to the right and back to
the left (assuming the default bottom dock widget area), and therefore
causes flickering.

bc4f29cb9ec65c82360daf14dd0b8aa6268c0fd6 introduced the parenthesizing
feature to indicate whether a tool view widget is disabled. Use the
inactive color for the text of a tool button instead of parenthesizing
it in order to get rid of the resizing and the flickering.