test_lldb: get rid of verifyStackFrameCountNotLessThan() to simplify
2b45f1ce42218e046d8ab91267a1d5e34041a509 introduced
verifyStackFrameCountNotLessThan() to fix LldbTest::testStack*() back
when KDevelop depended on Qt 5. Replace all uses of
verifyStackFrameCountNotLessThan() with the macro QCOMPARE_GE()
introduced in Qt 6.4 and remove the custom function to simplify.

Convert into a NOTE and move from the removed function to the scope of
an unnamed namespace the comment that explains why the stack frame count
is not verified via QCOMPARE().

Expand the comment with the names of the frames at the bottom of a call
stack in the current version of my GNU/Linux system that has LLDB
version 20.1.8:
    main
    __libc_start_call_main
    __libc_start_main_impl
    _start
The message of the aforementioned commit explains on which systems the
other frame names are reported.