Extract KDevMI::Testing::TestFrameStackModel
test_gdb and test_lldb define practically identical classes
TestFrameStackModel that differ only in the base class and in the type
of the debug session. So extract a class template to deduplicate.

A class template cannot contain the Q_OBJECT macro, but it is not needed
in the class TestFrameStackModel, so removing the macro in the course of
the deduplication should be fine.