Debugger tests: decrease debuggee recursion depth to simplify test code
The total number of stack frames fetched by KDevelop::FrameStackModel
grows as follows: 21, 41, 121, 301. testStackFetchMore() of GdbTest and
LldbTest tests consecutive frame counts of debugeerecursion, which
invokes foo() recursively to the depth of 298. The first call to foo()
is from main(), so the maximum stack frame count in GdbTest is 299,
which is less than a growth step 301.

LldbTest::testStackFetchMore() has been copied from GdbTest and
adjusted slightly to account for the 2 extra frames reported by
kdevlldb. 299+2=301 still fits into the growth step. But at some point
kdevlldb started reporting 3 instead of 2 extra frames. This caused a
failure of LldbTest::testStackFetchMore() because 302>301, which made
the comment "//nothing to fetch, we are at the end" wrong, and set
stackModel->fetchFramesCalled to 5 rather than the expected 4 and
stackModel->rowCount(tIdx) to 302 instead of the expected 301.

2b45f1ce42218e046d8ab91267a1d5e34041a509 fixed the test failure by
verifying the minimum expected frame count of 299 or more. But that
commit introduced a loop into LldbTest::testStackFetchMore(), which
complicates the test code and makes extracting testStackFetchMore() into
DebuggerTestBase to deduplicate less straightforward. Reduce the
recursion depth of debugeerecursion from 298 to 295 and (mostly) revert
to the previous simpler version of LldbTest::testStackFetchMore().