Extract DebuggerTestBase::testStack*()
The 3 slots testStack*() of GdbTest and LldbTest are almost identical,
so extract them into the common base class to deduplicate.

Use ActiveStateSessionSpy and the safer macros that take the spy as an
argument in the extracted test functions because the spy is necessary to
prevent flakiness of test_lldb and does not harm test_gdb.

33046ef749de05627324b3c99a32646206017edf made
GdbTest::testStackSwitchThread() more resilient by replacing a
QTest::qWait() and a QVERIFY() with a QTRY_VERIFY(). In contrast,
LldbTest::testStackSwitchThread() still waits 200 ms and then calls
QCOMPARE_GE(). Keep QCOMPARE_GE() for better output on failure. Wait for
the paused state and idle instead of the unreliable 200 ms wait and the
less reliable (because it can detect a hypothetical intermediate state
with enough stack frames) QTRY_VERIFY() approach. A debugger sends
frames in the reply to the MI command -stack-list-frames that is issued
during the call to MIFrameStackModel::setCurrentThread(), so the
expected frames are guaranteed to be in the stack model when the
debugger becomes idle.

Extract VALIDATE_COLUMN_COUNTS_THREAD_COUNT_AND_STACK_FRAME_NUMBERS(),
use it in place of validateColumnCountsThreadCountAndStackFrameNumbers()
and RETURN_IF_TEST_FAILED() to make test code more concise and readable.

Apply trivial cleanups to the deduplicated code:
* use more `const` and `auto` where possible;
* rename local variables from tIdx to threadIndex and from last to
  currentLocation for clarity;
* use the more readable helper function addDebugeeBreakpoint().