Extract DebuggerTestBase::testStack*()
The 3 member functions testStack*() of GdbTest and LldbTest are almost
identical, so extract them into the common base class to deduplicate.

Use ActiveStateSessionSpy and the safer macros that take the spy as an
argument in the extracted test functions because the spy is necessary to
prevent flakiness of test_lldb and does not harm test_gdb.

378f03ee8c1c5d423fb0ca87923fbb7a86d02a4b worked around a failure of
GdbTest::testStack() on the CI by allowing the previous line to be
current after stepping out (i.e. after the MI command -exec-finish). The
test failure might have been caused by some transient bug, and test_gdb
does not need to tolerate such surprising behavior forever. So remove
the workaround.

33046ef749de05627324b3c99a32646206017edf made
GdbTest::testStackSwitchThread() more resilient by replacing a
QTest::qWait() and a QVERIFY() with a QTRY_VERIFY(). A few years later
ec391e500bfd66a024e2622f210097bb09c8a820 inserted a call to
validateColumnCountsThreadCountAndStackFrameNumbers() between the
modifying function call stackModel->setCurrentThread(2) and the waiting-
and-verifying QTRY_VERIFY() call. Unlike the only-verifying QCOMPARE()
calls, the order of the waiting and the validation is important, so
reorder the validation after the waiting. Use the more robust frame
count verification of GdbTest, but replace the QTRY_VERIFY() call with
an equivalent QTRY_COMPARE_GE() call to improve output on failure.

Extract VALIDATE_COLUMN_COUNTS_THREAD_COUNT_AND_STACK_FRAME_NUMBERS(),
use it in place of validateColumnCountsThreadCountAndStackFrameNumbers()
and RETURN_IF_TEST_FAILED() to make test code more concise and readable.

Apply trivial cleanups to the deduplicated code:
* use more `const` and `auto` where possible;
* rename local variables from tIdx to threadIndex for clarity;
* use the more readable helper function addDebugeeBreakpoint().