Extract and improve MIDebugSession::setCurrentPositionToFrameFieldOf()
The extracted function will be reused in the next commit.

Extract literalIfHasField() from the body of the extracted function into
the unnamed namespace. literalIfHasField() is general enough to be
reused in another translation unit, so it can be moved into some header
if and when needed.

Refactor away a superfluous local variable updateState.