Set current position while examining a core file
The position of the crash is usually present in a core file. But it is
not displayed in some of the KDevelop UI when a core file is loaded,
because IDebugSession::setCurrentPosition() is not invoked.

Execute a -stack-info-frame MI command and pass its result to
IDebugSession::setCurrentPosition() in order to correct this omission.

Note that a debugger sends the same current position information in the
replies to the already issued from elsewhere MI commands -thread-info
and -stack-info-frame. However, the handlers of the results of these
commands do not always store the current address (the value of the
"addr" field). A core file is loaded relatively rarely, so executing an
extra -stack-info-frame command is preferable to the complication of
retrieving the address from the handlers for the thread and frame info
commands that are issued in more scenarios and more often.

Now the following features work while examining a core file:
1. An execution point is set (i.e. the Execution text document mark is
   displayed) at the location of the crash.
2. The action Show Current Line activates the location of the crash. An
   auxiliary change in this commit is that now the action Show Current
   Line is enabled rather than disabled in the stopped state.
3. The Disassemble/Registers tool view can work. Though only the
   Registers view actually works right now because DisassembleWidget
   returns early instead of issuing MI commands if the current debug
   session returns false from IDebugSession::isRunning(). isRunning()
   returns true only if the session is in the active or the paused
   state, but the state is stopped while a core file is being examined.
   A subsequent commit will remove this restriction and make the
   Disassemble view work too.