Make public, rename and reuse RunController::slotDebug()
DebugController::run() inlines the definition of the slot
RunController::slotDebug(), except that it does not check the list of
launch configurations for emptiness after showing the configuration
dialog. Therefore, if the list is still empty after the user dismisses
the launch configuration dialog opened by DebugController::run(),
RunController::executeDefaultLaunch() is called, cannot execute a
program, and so just prints a warning "no default launch!".

Reuse the slot RunController::slotDebug() in DebugController::run() to
eliminate the duplication and the warning. The slot is connected only
via the function pointer-based syntax, so convert it into a simple
member function.