DebugController: make Continue/Debug action work in stopped state
When a core file is loaded, the MI debug session enters the stopped
state. Continuing the application execution is impossible, so
DebugController::updateDebuggerState() converts the Continue action into
the Debug Launch action. However, triggering this Debug Launch action
has no effect because DebugController::run() starts a new debug session
only if there is no current session.

Store the current mode of the Continue/Debug action in a new data member
bool DebugController::m_continueStartsDebug and check it instead of the
current session pointer in DebugController::run(). Now triggering the
Continue/Debug action in the stopped state asks the user whether to
abort the currently running debug session and continue with the (new)
launch - just like triggering the other, unchanging Debug Launch action
of RunController does.

DebugController::setupActions() calls
DebugController::setContinueStartsDebug(true), so the newly created
action starts in the Debug Launch mode. It becomes a Continue action
only when a current debug session enters the starting, paused or active
state. The current session can become null only in
DebugController::debuggerStateChanged(). debuggerStateChanged() calls
updateDebuggerState(EndedState) before setting the action to null, which
converts the Continue action back into the Debug Launch action.
Therefore, we can safely assume and assert that the current session is
not null when the action becomes a Continue action and when the Continue
action is triggered.

Optimize by returning early from setContinueStartsDebug() if the action
is already in the requested mode. Initialize m_continueStartsDebug to
false to make the setContinueStartsDebug(true) call in setupActions()
initialize the UI properties of the action when it is created.