Extract DebuggerTestBase::testCoreFile()
Deduplicate by extracting into the common base class the slots named
testCoreFile() of GdbTest and LldbTest. The definitions of the two slots
differ only in an extra debug output line in LldbTest's version (copy it
to the extracted slot and remove a comment just above the debug output,
which repeats the same message) and in an expected stack model frame
name (use DebuggerTestBase::adjustedStackModelFrameName() to deal with
this difference between the outputs of GDB/MI and LLDB-MI).

Apply trivial cleanups to the deduplicated code:
* use more `const` and `auto` where possible;
* rename a local variable from tIdx to threadIndex for clarity;
* remove all uses of QStringLiteral() to improve readability.