Terminate LLDB debug session if loading core file fails
42fa01d4023d49f40d54dd39cdcd874d5af39936 started looking for error
messages in the console stream output produced while LLDB loads a core
file. Presumably, then as now, unlike GDB/MI, LLDB-MI always replied
with "done" to the command that loads a core file, even if an error has
occurred and no core file has been loaded. Now on my GNU/Linux system,
when LLDB version 20.1.8 fails to load a core file, it prints an error
message to the *log* stream output where
LLDB::DebugSession::handleCoreFile() does not look for it. So the debug
session is stuck in the stopped state instead of ending with an error.

Store the log stream output along with the console stream output in the
LLDB core-file-loading MI command object and look for the error message
in the combined output. Now an LLDB debug session correctly ends with an
error when loading a core file fails, just like a GDB debug session.

The added test data rows reveal that a GDB debug session is stuck in the
starting state if the core filename is empty. Use QEXPECT_FAIL() to
avoid introducing a failure of test_gdb.