Move the beginning of loadCoreFile() to MIDebugSession
The definitions of the member functions loadCoreFile() of
GDB::DebugSession and LLDB::DebugSession begin with two identical lines
of code. Move the two lines before the only call to loadCoreFile() in
MIDebugSession::examineCoreFile() to deduplicate. Adjust the
documentation of the pure virtual function
MIDebugSession::loadCoreFile() accordingly.

Remove two unused parameters and return void instead of always-true bool
from loadCoreFile(). Switch the name of the remaining parameter to the
conventional camel case.