DisassembleWindow: disable inapplicable actions in stopped state
As the parent commit message explains, the actions Jump to Cursor and
Run to Cursor can work only if the current MI debug session is running,
i.e. is in the active or the paused state. So disable these actions if
the session is not running to reflect this requirement in the UI.