MemoryView: don't issue MI commands that inevitably fail
If the inferior (app) has not started, MemoryView disables the OK button
of the range selector and the context menu actions, triggering which
issues MI commands. However, if the debugger state changes before the MI
command is issued (e.g. while the context menu is visible), MemoryView
can theoretically issue an MI command at a wrong time. Issue the MI
commands -data-evaluate-expression and -data-read-memory only when the
inferior (app) has started (mostly for consistency).

While a core file is being examined:
* evaluating expressions and reading memory works correctly;
* but modifying memory contents in the memory view and triggering the
  context menu action Write Changes issues the MI command -gdb-set,
  which fails and has no effect (GDB replies with an error message
  "Cannot access memory at address 0x7ffc52b53b0f").

The user is likely to miss the failure to write memory changes because
the memory view contents stays edited and is marked as unmodified - just
as when writing to memory succeeds. The GDB error message does appear in
the GDB tool view, but only if its context menu checkbox Show Internal
Commands is checked (it is unchecked by default).

So make use of the fact that the current session is in the stopped state
while a core file is being examined to prevent the misleading behavior.
Namely, enable the context menu action Write Changes and issue the MI
command -gdb-set only if the current session is running, i.e. is in the
active or the paused state.