test_gdb: don't wait fixed intervals before or after paused&idle
The fixed-interval waits removed in this commit were introduced back
when WAIT_FOR_STATE(session, DebugSession::PausedState) did not wait for
the idle state of the debugger. Nowadays the extra waits are unneeded as
evidenced by the still-passing test functions. The corresponding test
functions of test_lldb do not contain any of the removed waits either.