test_gdb: don't wait for idle between adding catchpoint and continuing
3c286c439df126a7879705800ee691d2207b4b8d added this waiting apparently
to make the test non-flaky, with the following brief explanation:
    fix race-condition, wait for "catch throw" done

But seeing as the next command is sent only after the debugger replies
to the previous one, and that a breakpoint is added right when GDB
replies to the "catch throw" command, the waiting should not be needed.
There is no such waiting after non-MI commands in other test functions.
The corresponding test function of test_lldb does not contain such a
wait either.

I have no idea why the extra waiting had made a difference back then
(and I am not even sure it did make a difference). Maybe the test had
failed due to some long-since fixed bug. Our unit tests do not need to
support every short-lived GDB bug, so remove the redundant waiting.