test_gdb: unite 3 testRemoteDebug* functions into one - data-driven
The definitions of GdbTest::testRemoteDebug*() are very similar, so the
3 functions can be deduplicated into a single data-driven test function.

Out of the 3 functions, only
testRemoteDebugInsertBreakpointPickupOnlyOnce() converts file paths to
QByteArray via QString::toLatin1() rather than via QString::toUtf8().
Use the safer toUtf8() in the united function.

Capture the differences in added debugee breakpoints and in starting (or
not starting) a second debug session via two test data parameters. The
order of the 3 test data rows matches the order of the corresponding
former test functions.

Verify that the number of breakpoints is 1 in the first data row, which
corresponds to the original function testRemoteDebug() that missed the
opportunity to check that number.

Apply trivial cleanups to the deduplicated code:
* use the more readable helper functions addDebugeeBreakpoint() and
  currentMiLine();
* use QList::size() in place of QList::count() for compatibility with
  the standard library name.