DebuggerTestBase: verify that debugger stops at regex breakpoints
Both GDB and LLDB stop at the expected line numbers 20 and 24 in
DebuggerTestBase::testRegularExpressionBreakpoint(). Verify this
expected behavior instead of removing the breakpoints.

Note that both in testMultipleLocationsBreakpoint() and in
testRegularExpressionBreakpoint() KDevelop saves only a single simple
code breakpoint at the line 24 after an LLDB debug session. But during
the session in which a multiple-location or a regular-expression
breakpoint is added, that breakpoint works correctly despite being
misrepresented in the KDevelop UI.