DebuggerTestBase: don't wait fruitlessly after switching frame
Setting the current frame is handled synchronously and does not cause
any MI commands to be sent if automatic local variable updates are not
enabled and there are zero variable watches, as is the case in
DebuggerTestBase::testSwitchFrameDebuggerConsole(). So no need to wait
half a second and verify that the current frame does not change during
that time in this test function.