Debugger tests: wait for state&idle after adding a watch
Whenever one of the functions testVariables*() adds a watch, it waits
300 ms (and in one case 3000 ms). The likely purpose of the wait is to
ensure that all pending commands have been sent and replied to by the
debugger, after which the added watch has the expected value that can be
verified. Achieve the same result more reliably by waiting for the idle
state of the debugger. Waiting for the idle state should also reduce the
average running time of the test functions.