GdbTest: don't use QTRY_COMPARE() needlessly.
519637dc4d9ca356f7b39283774a11f20896067d replaced a QCOMPARE() statement
with QTRY_COMPARE() in GdbTest::testPickupCatchThrowOnlyOnce(). The
message explains the other uniform changes in that commit, but not the
addition of TRY_. The comparison occurs after a debug session has ended,
and the number of breakpoints in the model cannot change on its own
afterwards.

Restore the original QCOMPARE() statement. Move it into the loop to also
verify that there is a single breakpoint when the first debug session
ends.

Expand a comment to better explain the idea of the comparison and of the
entire test function.