Make DebuggerTestBase::testIgnoreHitsBreakpoint() work
3f9b8a350b48444c1c073e07c7d7a2ffbf11ad71 introduced a new class
TestWaiter and a wrapper macro WAIT_FOR(). The new macro became used in
GdbTest::testIgnoreHitsBreakpoint() in place of waiting for the paused
state plus 100 ms so as to reliably skip the intermediate paused state
during program startup and thus increase robustness.

However, the author of that commit apparently misunderstood the relation
between the hit count and the ignore hits properties of a breakpoint. In
fact, the hit count is incremented even if a breakpoint hit is ignored.
Consequently, the second WAIT_FOR() call returns immediately without
waiting and the test function does not work as intended.

The grandchild of the aforementioned commit by the same author,
e1279c961b8b3076f52b4def132850a7d366b2d2, introduced a more general and
convenient way to skip the intermediate paused state during program
startup - WAIT_FOR_STATE_AND_IDLE(). Use it in place of WAIT_FOR() to
reveal the wrong hit count assumptions. Also use ActiveStateSessionSpy
to make the test function work reliably in test_lldb as well (test_lldb
skips testIgnoreHitsBreakpoint() at this time, but it does not have to).
Remove the no longer used class TestWaiter and macro WAIT_FOR().

The debugee line 22 of the first breakpoint is no-op, so a debugger
moves it to the next line 23. The fact that 23 is also the line of the
second breakpoint impedes checking which breakpoint is hit. Increment
the line numbers of the two breakpoints to get rid of this obstacle.
Check the line at which the debugger stops during each paused state to
verify that the expected breakpoint is hit.

Verify the hit counts of both breakpoints at each stage of the debug
session.