DebuggerTestBase: wait for state&idle after changing breakpoint location
DebuggerTestBase::testChangeLocationBreakpoint() changes the location of
an existing breakpoint, then waits 100 ms and verifies that the line
number of the breakpoint has not been reverted or modified again. The
most likely intention is to wait for the debugger to send back the
updated breakpoint and verify that KDevelop does not modify the
breakpoint's line because of it. So wait for the paused state and idle
instead of the unreliable fixed interval.

Both GDB/MI and LLDB-MI send the updated breakpoint record in the reply
to the MI command -break-insert that is issued during the call to
BreakpointModel::setData(). Consequently, this breakpoint update is
guaranteed to be processed before the debugger becomes idle. LLDB-MI
also sends exactly the same breakpoint info in two breakpoint-modified
notifications immediately after the reply to the MI command. These
notifications are processed after the debugger becomes idle, but the
added call to WAIT_FOR_STATE_AND_IDLE() is still unlikely to return
before the processing. If it does, occasionally verifying the absence of
a breakpoint change less thoroughly should not be a problem.