DocumentChangeTracker: remove no longer needed members
0931fe88180d18d03edc418b97e805f4a71a7178 removed the definition but not
the declaration of DocumentChangeTracker::checkMergeTokens(). That
commit also removed all uses of a commented out macro ALWAYS_UPDATE, but
failed to remove the macro itself.

The same commit also made DocumentChangeTracker::m_needUpdate and
DocumentChangeTracker::needUpdate() unnecessary, but failed to simplify
the code. The simplification is possible because the same expression is
unconditionally assigned to m_needUpdate just before every call to
DocumentChangeTracker::updateChangedRange(), which is the only caller of
needUpdate(), which in turn simply returns m_needUpdate.