IVariableController: don't update variables repeatedly
If the Variables tool view is visible and at least one tree item within
(Auto or Locals) is expanded, IVariableController::handleEvent() usually
updates variables twice when its debug session enters the paused state
(except for the very first paused state of the session).

The two variable updates occur because KDevelop::FrameStackModel raises
the event thread_or_frame_changed multiple times and IVariableController
observes two different values of IFrameStackModel::currentFrame(). The
first current frame value -1 is observed when
FrameStackModel::stateChanged() calls setCurrentFrame(-1). The second
value 0 is observed when FrameStackModel::setCurrentThread() or
FrameStackModel::setFrames() sets FrameStackModelPrivate::m_currentFrame
to 0.

A single current frame value 0 is observed when the debug session
becomes paused for the first time, because
FrameStackModelPrivate::m_currentFrame is initialized to -1. So when
setCurrentFrame(-1) is called for the first time, the current frame is
unchanged, and therefore the event thread_or_frame_changed is not
raised. Later, FrameStackModel::setCurrentThread() simultaneously sets
the current thread to a positive value and the current frame to 0, which
triggers a single update of variables.

The practical downsides of the redundant variable updates:
1. Just updated variables are not highlighted with the NeutralText
   (orange) color as intended, because setCurrentFrame(-1) is invoked
   and consequently the first MI command -var-update is queued before
   the MI commands -thread-info and -stack-list-frames. So the handler
   function for -var-update MIVariableController::handleVarUpdate()
   marks updated variables as changed. Then all variables are marked as
   unchanged again during the second invocation of
   IVariableController::updateIfFrameOrThreadChanged(). And finally the
   second invocation of handleVarUpdate() does not mark any variable as
   changed, because none has been updated since the previous invocation.
2. Issuing useless duplicate MI commands can noticeably slow down
   debugging, especially because performing operations on MI variable
   objects can take arbitrarily long time (depending on the debuggee and
   its input).

Return early from IVariableController::handleEvent() if the current
thread or frame is invalid (negative) to prevent the redundant variable
updates. Also do not update variables in
IVariableController::setAutoUpdate() if the current thread or frame
cached in IVariableControllerPrivate is invalid (this scenario is
difficult to reproduce).

Soon after the skipping of the update, the current thread and frame will
probably become valid, at which point handleEvent() will update
variables. Updating variables in an invalid frame makes no sense anyway,
so if the current thread and frame do not become valid eventually, not
updating variables at all is correct.

Even with the fixes in this commit, just updated variables are still not
highlighted in kdevlldb as a comment at the top of the added test
function explains. Without the fixes in this commit, test_gdb fails in
the two places where test_lldb xfails within the added test function.

Two alternative fixes have been considered and rejected because they
only reduce the frequency of duplicate variable updates but not
eliminate them, and because they can cause bugs in the Frame Stack tool
view (see e.g. 5fdbb0d1bf5febaf7497f0ea947ff82fc7c6baaf):
1. Replace setCurrentFrame(-1) with setCurrentFrame(0). The problem with
   this fix is that setCurrentFrame(0) still raises the event
   thread_or_frame_changed if
   FrameStackModelPrivate::m_currentFrame != 0 before the call.
   IVariableController would observe two different current
   (thread, frame) pairs when its debug session becomes paused for the
   first time: (-1, 0) and (T, 0) where T > 0. Also if the current frame
   is set to a positive value during a paused state, and the current
   thread (T > 0) changes at the next paused state (to U > 0, U != T),
   IVariableController would observe the following current
   (thread, frame) pairs: (T, 0) and (U, 0).
2. Simply remove setCurrentFrame(-1). The problem with this fix is again
   that 0 is merely the most common value of
   FrameStackModelPrivate::m_currentFrame when the debug session enters
   the paused state. If the current frame is set to a positive value
   (F > 0) during a paused state, and the current thread (T > 0) does
   not change at the next paused state, IVariableController would
   observe the following current (thread, frame) pairs: [set from
   FrameStackModel::setCurrentThread()] (T, F) and [set from
   FrameStackModel::setFrames()] (T, 0).