Correct a mistake in the documentation for Path::isParentOf()
715d09e5545758af0ba85c8fea7c11ecb94eb9ad replaced KUrl with QUrl in the
note within the documentation for Path::isParentOf() while porting from
KUrl to QUrl. KUrl::isParentOf(const KUrl& u) is documented to return
"true if this url is a parent of u (or the same URL as u)", so the note
with KUrl had been correct. However, QUrl::isParentOf() returns false if
the argument equals the object on which the member function is called.
Adjust the note in the Path documentation accordingly.