test_path: adapt to a change of QUrl::adjusted() in Qt 6.8.1
Since Qt 6.8.1, the following test_path failures occur regularly.
1. On GNU/Linux and FreeBSD:
FAIL!  : TestPath::testPath(clean-path) Compared values are not the same
   Actual   (optUrl.toUrl()): file:///asdf
   Expected (url)           : file:////asdf
   Loc: [kdevelop/kdevplatform/util/tests/test_path.cpp(223)]
2. On Windows:
FAIL!  : TestPath::testPath(clean-path) Compared values are not the same
   Actual   (optUrl.toUrl()): file:///C:/asdf
   Expected (url)           : file:///C://asdf
kdevelop\kdevplatform\util\tests\test_path.cpp(223) : failure location

The failures are due to a deliberate change in the behavior of
QUrl::adjusted(NormalizePathSegments). The behavior of the tested class
Path has not changed, but TestPath::testPath() computes expected values
using QUrl::adjusted(), and so these values changed in Qt 6.8.1. Remove
a possible duplicate slash from the path of a QUrl to restore the former
expected values with newer Qt versions.

Add a test data row for testPath() with a path that contains `cd up` out
of a non-root directory. Such a path is correctly verified even without
the adjustment implemented in this commit.