ci: don't run benchmarks
KDevelop's top-level CMakeLists.txt disables building and running
benchmarks if the build type is "Debug" or empty. The CMake variable
CMAKE_BUILD_TYPE is set to "Debug" on the OpenSUSE and FreeBSD CI
servers, but to "Release" on the Windows CI server. Consequently, all
benchmarks are built and run on Windows. Some of the benchmarks crash,
others occasionally time out.

Keep building benchmarks, at least on Windows, in order to verify that
developers, who often test only debug builds, do not accidentally break
them. But running the benchmarks during continuous integration merely
wastes CPU time and delays CI results. So make use of the fact that the
names of all KDevelop's benchmark executables consistently start with
"bench_" and tell ctest not to run them on CI servers.