lldb: refactor out ExecRunHandler autoDelete() override
This is in preparation for the removal of the autoDelete() method in a
later commit, which will prevent reusing a single handler instance to
handle results of multiple commands. Each command will have a unique
ownership of the assigned handler, and the handler instance will be
destroyed immediately after it is invoked.

Therefore, to remove the autoDelete() override, create a new
ExecRunHandler instance with a decremented m_remainRetry when the
ExecRun command fails. Allow the command to be retried at most five
times in this way, as before.

The handler will now be automatically destroyed because the default
implementation of MICommandHandler::autoDelete() returns true.
Consequently, the handler instances are cleaned up by either
MICommand::invokeHandler() or by the MICommand destructor. So remove the
no longer needed `delete this`.