Never leak FetchMoreChildrenHandler instances
FetchMoreChildrenHandler::autoDelete() returns false, which prevents
MICommand from destroying the handler. If a QPointer
FetchMoreChildrenHandler::m_variable happens to be null when
FetchMoreChildrenHandler::handle() is invoked,
FetchMoreChildrenHandler::m_activeCommands never reaches 0 and
consequently the FetchMoreChildrenHandler object is never destroyed.

MICommandHandler::autoDelete() will be removed in a subsequent commit,
which will prevent reusing a single handler instance to handle results
of multiple commands. Each command will have a unique ownership of the
assigned handler, and the handler instance will be destroyed immediately
after it is invoked.

In order to remove the autoDelete() override, clone the current instance
of FetchMoreChildrenHandler to handle the commands queued for access
specifier pseudo children.

The children of an access specifier pseudo child must be actual data
members of a struct or a class. The fix implemented in the parent commit
should prevent misdetection and thus guarantee that further commands are
never queued for such children. So the maximum recursion depth is 1, and
therefore the counter m_activeCommands can be replaced with a simpler
flag m_isLastHandler.

The new flag m_isLastHandler is initialized to true. When a pseudo child
is detected, the flag of the previous handler is set to false. The value
of the flag is checked after the pseudo child detection loop, and the
commands queued for pseudo children cannot run until the current handler
returns, so the flag is always updated before use. Therefore,
TreeItem::emitAllChildrenFetched() is called once, when the last command
in the series is handled, just like before this commit.

Each handler object is now destroyed automatically because the default
implementation of MICommandHandler::autoDelete() returns true.
Consequently, the handler instances are cleaned up by either
MICommand::invokeHandler() or by the MICommand destructor. So remove the
no longer needed `delete this`.