Remove leak-prone API MICommandHandler::autoDelete()
If a MICommand object is destroyed without a call to
MICommand::invokeHandler(), e.g. when
MI::CommandQueue::rationalizeQueue() removes a command, and the
command's handler returns false from autoDelete(), the handler instance
is not destroyed in ~MICommand() and is likely to be leaked.

The two previous commits removed two leak-prone overrides of
MICommandHandler::autoDelete(). Now only TestCommandHandler overrides
this function, and solely to test that the overriding works correctly.
Therefore, remove the problematic function
MICommandHander::autoDelete() and its single use in test_micommand.

Replace the raw pointer type of MICommand::commandHandler_ with
std::unique_ptr in order to further simplify and improve the safety of
the lifetime management of MICommandHandler. Rename commandHandler_ to
m_commandHandler so as to match the prevalent code style of KDevelop.

The type of m_commandHandler could be std::shared_ptr instead. This
alternative design was rejected for the following reasons:
1. It is not expected that a future command handler would meaningfully
   benefit from handling multiple commands by a shared handler instance.
   Not enough to justify the code complexity increase due to the use of
   std::shared_ptr anyway.
2. Higher overhead than std::unique_ptr due to the atomic counter.
3. With std::unique_ptr, MICommand is the sole owner of its handler
   instance. This provides a stricter guarantee of the handler's
   destruction and aligns with the fact that there is no API to get or
   take the handler from a MICommand.

Apply trivial code cleanups to close-by code of test_micommand:
* make TestCommandHandler::m_expectedRecordsHandled const;
* fix the order of arguments to QCOMPARE() in ~TestCommandHandler().